+++ src/chroot/chroot.rs
@@ -40,7 +40,7 @@ extern {
 static NAME: &'static str = "chroot";
 static VERSION: &'static str = "1.0.0";
 
-pub fn uumain(args: Vec<String>) -> isize {
+pub fn uumain(args: Vec<String>) -> i32 {
     let program = &args[0];
 
     let options = [
@@ -98,7 +98,7 @@ pub fn uumain(args: Vec<String>) -> isize {
         let executable = CString::from_slice(command[0].as_bytes()).as_bytes_with_nul().as_ptr();
         let mut command_parts: Vec<*const u8> = command.iter().map(|x| CString::from_slice(x.as_bytes()).as_bytes_with_nul().as_ptr()).collect();
         command_parts.push(std::ptr::null());
-        execvp(executable as *const libc::c_char, command_parts.as_ptr() as *mut *const libc::c_char) as isize
+        execvp(executable as *const libc::c_char, command_parts.as_ptr() as *mut *const libc::c_char) as i32 
     }
 }
