+++ src/paste/paste.rs
@@ -23,7 +23,7 @@ mod util;
 static NAME: &'static str = "paste";
 static VERSION: &'static str = "1.0.0";
 
-pub fn uumain(args: Vec<String>) -> isize {
+pub fn uumain(args: Vec<String>) -> i32 {
     let program = args[0].clone();
 
     let opts = [
